requirements.txt
beautifulsoup4==4.13.4
openai==1.82.0
PyYAML==6.0.2
requests==2.32.3
